requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
python-dotenv>=1.0.0
requests>=2.31.0
jinja2>=3.1.0
pandas>=2.0.0
numpy>=1.24.0
python-multipart>=0.0.6
aiosqlite>=0.19.0
scikit-learn>=1.3.0
joblib>=1.3.0
